+++ lib/utilcode/src/main/java/com/blankj/utilcode/util/ActivityUtils.java
@@ -1645,15 +1645,14 @@ public final class ActivityUtils {
                                            final boolean isIncludeSelf,
                                            final boolean isLoadAnim) {
         List<Activity> activities = UtilsBridge.getActivityList();
-        for (int i = activities.size() - 1; i >= 0; --i) {
-            Activity aActivity = activities.get(i);
-            if (aActivity.equals(activity)) {
+        for (Activity act : activities) {
+            if (act.equals(activity)) {
                 if (isIncludeSelf) {
-                    finishActivity(aActivity, isLoadAnim);
+                    finishActivity(act, isLoadAnim);
                 }
                 return true;
             }
-            finishActivity(aActivity, isLoadAnim);
+            finishActivity(act, isLoadAnim);
         }
         return false;
     }
@@ -1673,15 +1672,14 @@ public final class ActivityUtils {
                                            @AnimRes final int enterAnim,
                                            @AnimRes final int exitAnim) {
         List<Activity> activities = UtilsBridge.getActivityList();
-        for (int i = activities.size() - 1; i >= 0; --i) {
-            Activity aActivity = activities.get(i);
-            if (aActivity.equals(activity)) {
+        for (Activity act : activities) {
+            if (act.equals(activity)) {
                 if (isIncludeSelf) {
-                    finishActivity(aActivity, enterAnim, exitAnim);
+                    finishActivity(act, enterAnim, exitAnim);
                 }
                 return true;
             }
-            finishActivity(aActivity, enterAnim, exitAnim);
+            finishActivity(act, enterAnim, exitAnim);
         }
         return false;
     }
@@ -1708,15 +1706,14 @@ public final class ActivityUtils {
                                            final boolean isIncludeSelf,
                                            final boolean isLoadAnim) {
         List<Activity> activities = UtilsBridge.getActivityList();
-        for (int i = activities.size() - 1; i >= 0; --i) {
-            Activity aActivity = activities.get(i);
-            if (aActivity.getClass().equals(clz)) {
+        for (Activity act : activities) {
+            if (act.getClass().equals(clz)) {
                 if (isIncludeSelf) {
-                    finishActivity(aActivity, isLoadAnim);
+                    finishActivity(act, isLoadAnim);
                 }
                 return true;
             }
-            finishActivity(aActivity, isLoadAnim);
+            finishActivity(act, isLoadAnim);
         }
         return false;
     }
@@ -1736,15 +1733,14 @@ public final class ActivityUtils {
                                            @AnimRes final int enterAnim,
                                            @AnimRes final int exitAnim) {
         List<Activity> activities = UtilsBridge.getActivityList();
-        for (int i = activities.size() - 1; i >= 0; --i) {
-            Activity aActivity = activities.get(i);
-            if (aActivity.getClass().equals(clz)) {
+        for (Activity act : activities) {
+            if (act.getClass().equals(clz)) {
                 if (isIncludeSelf) {
-                    finishActivity(aActivity, enterAnim, exitAnim);
+                    finishActivity(act, enterAnim, exitAnim);
                 }
                 return true;
             }
-            finishActivity(aActivity, enterAnim, exitAnim);
+            finishActivity(act, enterAnim, exitAnim);
         }
         return false;
     }
@@ -1768,10 +1764,9 @@ public final class ActivityUtils {
     public static void finishOtherActivities(@NonNull final Class<? extends Activity> clz,
                                              final boolean isLoadAnim) {
         List<Activity> activities = UtilsBridge.getActivityList();
-        for (int i = activities.size() - 1; i >= 0; i--) {
-            Activity activity = activities.get(i);
-            if (!activity.getClass().equals(clz)) {
-                finishActivity(activity, isLoadAnim);
+        for (Activity act : activities) {
+            if (!act.getClass().equals(clz)) {
+                finishActivity(act, isLoadAnim);
             }
         }
     }
@@ -1789,10 +1784,9 @@ public final class ActivityUtils {
                                              @AnimRes final int enterAnim,
                                              @AnimRes final int exitAnim) {
         List<Activity> activities = UtilsBridge.getActivityList();
-        for (int i = activities.size() - 1; i >= 0; i--) {
-            Activity activity = activities.get(i);
-            if (!activity.getClass().equals(clz)) {
-                finishActivity(activity, enterAnim, exitAnim);
+        for (Activity act : activities) {
+            if (!act.getClass().equals(clz)) {
+                finishActivity(act, enterAnim, exitAnim);
             }
         }
     }
@@ -1811,12 +1805,11 @@ public final class ActivityUtils {
      */
     public static void finishAllActivities(final boolean isLoadAnim) {
         List<Activity> activityList = UtilsBridge.getActivityList();
-        for (int i = activityList.size() - 1; i >= 0; --i) {// remove from top
-            Activity activity = activityList.get(i);
+        for (Activity act : activityList) {
             // sActivityList remove the index activity at onActivityDestroyed
-            activity.finish();
+            act.finish();
             if (!isLoadAnim) {
-                activity.overridePendingTransition(0, 0);
+                act.overridePendingTransition(0, 0);
             }
         }
     }
@@ -1832,11 +1825,10 @@ public final class ActivityUtils {
     public static void finishAllActivities(@AnimRes final int enterAnim,
                                            @AnimRes final int exitAnim) {
         List<Activity> activityList = UtilsBridge.getActivityList();
-        for (int i = activityList.size() - 1; i >= 0; --i) {// remove from top
-            Activity activity = activityList.get(i);
+        for (Activity act : activityList) {
             // sActivityList remove the index activity at onActivityDestroyed
-            activity.finish();
-            activity.overridePendingTransition(enterAnim, exitAnim);
+            act.finish();
+            act.overridePendingTransition(enterAnim, exitAnim);
         }
     }
 
@@ -1854,7 +1846,7 @@ public final class ActivityUtils {
      */
     public static void finishAllActivitiesExceptNewest(final boolean isLoadAnim) {
         List<Activity> activities = UtilsBridge.getActivityList();
-        for (int i = activities.size() - 2; i >= 0; i--) {
+        for (int i = 1; i < activities.size() - 1; i++) {
             finishActivity(activities.get(i), isLoadAnim);
         }
     }
@@ -1870,7 +1862,7 @@ public final class ActivityUtils {
     public static void finishAllActivitiesExceptNewest(@AnimRes final int enterAnim,
                                                        @AnimRes final int exitAnim) {
         List<Activity> activities = UtilsBridge.getActivityList();
-        for (int i = activities.size() - 2; i >= 0; i--) {
+        for (int i = 1; i < activities.size() - 1; i++) {
             finishActivity(activities.get(i), enterAnim, exitAnim);
         }
     }